[DFT] Move classes to detail namespace to avoid symbol collisions with MKLGPU
* Move DFT classes to detail namespace to avoid symbol collisions with closed-source MKLGPU backend.
